First functional release of the CoreMetaDataService
This release is capable of creating an instance of the MetaData class
from the MetaModel instance reflected from a K2 Application
configuration.

The MetaData class implements The K2DaoFactory and supplies K2Dao
instances for the CoreMetaDataClasses populated with data drawn from all
the service classes reflected from the supplied K2 Application.